// src/lib/vanillaApi/models/SpecIdentifier.ts
// tslint:disable
/**
 * Vanilla API
 * This API helps calculate gear effectiveness
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum SpecIdentifier {
    FireMage = 'FireMage',
    FrostMage = 'FrostMage',
    ArcaneMage = 'ArcaneMage',
    Warlock = 'Warlock',
    BalanceDruid = 'BalanceDruid',
    ElementalShaman = 'ElementalShaman',
    ShadowPriest = 'ShadowPriest'
}

export function SpecIdentifierFromJSON(json: any): SpecIdentifier {
    return json as SpecIdentifier;
}

export function SpecIdentifierToJSON(value?: SpecIdentifier): any {
    return value as any;
}