// src/lib/vanillaApi/models/ReqFields.ts
// tslint:disable
/**
 * Vanilla API
 * This API helps calculate gear effectiveness
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    SpecIdentifier,
    SpecIdentifierFromJSON,
    SpecIdentifierToJSON,
    Stats,
    StatsFromJSON,
    StatsToJSON,
} from './';

/**
 * 
 * @export
 * @interface ReqFields
 */
export interface ReqFields {
    /**
     * 
     * @type {Stats}
     * @memberof ReqFields
     */
    stats: Stats;
    /**
     * 
     * @type {SpecIdentifier}
     * @memberof ReqFields
     */
    spec: SpecIdentifier;
}

export function ReqFieldsFromJSON(json: any): ReqFields {
    return {
        'stats': StatsFromJSON(json['stats']),
        'spec': SpecIdentifierFromJSON(json['spec']),
    };
}

export function ReqFieldsToJSON(value?: ReqFields): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'stats': StatsToJSON(value.stats),
        'spec': SpecIdentifierToJSON(value.spec),
    };
}